# web/src/assets/swagger/swagger-steambuddy.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "v1",
    "title" : "Steambuddy TypeScript API"
  },
  "tags" : [ {
    "name" : "FriendRequestResource"
  }, {
    "name" : "GameResource"
  }, {
    "name" : "GroupResource"
  }, {
    "name" : "UserResource"
  } ],
  "paths" : {
    "internal/api/friend-request/friend-requests/accept/{id}" : {
      "put" : {
        "tags" : [ "FriendRequestResource" ],
        "summary" : "Accept a friendRequest.",
        "description" : "Accept a friendRequest.",
        "operationId" : "acceptRequest",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        }
      }
    },
    "internal/api/friend-request/friend-requests/decline/{id}" : {
      "put" : {
        "tags" : [ "FriendRequestResource" ],
        "summary" : "Decline a friendRequest.",
        "description" : "Decline a friendRequest.",
        "operationId" : "declineRequest",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        }
      }
    },
    "internal/api/friend-request/friend-requests/incoming/{id}" : {
      "get" : {
        "tags" : [ "FriendRequestResource" ],
        "summary" : "Get all incomin friendRequests of a user.",
        "description" : "Returns all incoming friendRequests.",
        "operationId" : "getIncoming",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/FriendRequestDTO"
              }
            }
          }
        }
      }
    },
    "internal/api/friend-request/friend-requests/outgoing/{id}" : {
      "get" : {
        "tags" : [ "FriendRequestResource" ],
        "summary" : "Get all outgoing friendRequests of a user.",
        "description" : "Returns all outgoing friendRequests.",
        "operationId" : "getOutgoing",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/FriendRequestDTO"
              }
            }
          }
        }
      }
    },
    "internal/api/friend-request/friend-requests/{id}" : {
      "get" : {
        "tags" : [ "FriendRequestResource" ],
        "summary" : "Get friendRequest by id.",
        "description" : "Returns the friendRequest with given id.",
        "operationId" : "getRequestById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FriendRequestDTO"
            }
          }
        }
      }
    },
    "internal/api/friend-request/friend-requests/{sender_id}/{receiver_id}" : {
      "post" : {
        "tags" : [ "FriendRequestResource" ],
        "summary" : "Create a new friendRequest.",
        "description" : "Returns the id of the created friendRequest.",
        "operationId" : "createRequest",
        "parameters" : [ {
          "name" : "sender_id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "receiver_id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        }
      }
    },
    "internal/api/game/games" : {
      "get" : {
        "tags" : [ "GameResource" ],
        "summary" : "Get games which contain the given name",
        "description" : "Returns found games.",
        "operationId" : "getGamesByName",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/GameDTO"
              }
            }
          }
        }
      }
    },
    "internal/api/group/groups" : {
      "get" : {
        "tags" : [ "GroupResource" ],
        "summary" : "Get all groups.",
        "description" : "Returns all Groups.",
        "operationId" : "getGroups",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/GroupDTO"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "GroupResource" ],
        "summary" : "Create a new group.",
        "description" : "Returns the created Group.",
        "operationId" : "createGroup",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/GroupDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/GroupDTO"
            }
          }
        }
      }
    },
    "internal/api/group/groups/{user_id}" : {
      "get" : {
        "tags" : [ "GroupResource" ],
        "summary" : "Get all owned groups of user.",
        "description" : "Returns all owned Groups.",
        "operationId" : "getGroupsByUserId",
        "parameters" : [ {
          "name" : "user_id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/GroupDTO"
              }
            }
          }
        }
      }
    },
    "internal/api/group/{id}" : {
      "delete" : {
        "tags" : [ "GroupResource" ],
        "summary" : "Delete a Group.",
        "description" : "Delete a Group.",
        "operationId" : "deleteGroup",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/GroupDTO"
            }
          }
        }
      }
    },
    "internal/api/user/users" : {
      "get" : {
        "tags" : [ "UserResource" ],
        "summary" : "Get all users.",
        "description" : "Returns all Users.",
        "operationId" : "getUsers",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/UserDTO"
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "UserResource" ],
        "summary" : "Update a user.",
        "description" : "Returns the updated User.",
        "operationId" : "updateUser",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UserDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UserDTO"
            }
          }
        }
      }
    },
    "internal/api/user/users/authenticate" : {
      "put" : {
        "tags" : [ "UserResource" ],
        "summary" : "Authenticate a user.",
        "description" : "Returns the logged in User.",
        "operationId" : "authenticateUser",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UserDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UserDTO"
            }
          }
        }
      }
    },
    "internal/api/user/users/register" : {
      "post" : {
        "tags" : [ "UserResource" ],
        "summary" : "Register a new user.",
        "description" : "Returns the created User.",
        "operationId" : "registerUser",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UserDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UserDTO"
            }
          }
        }
      }
    },
    "internal/api/user/users/{id}" : {
      "get" : {
        "tags" : [ "UserResource" ],
        "summary" : "Get user by id.",
        "description" : "Returns user with given id.",
        "operationId" : "getUserById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UserDTO"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "FriendRequestDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "timestamp" : {
          "type" : "string",
          "format" : "date-time"
        },
        "accepted" : {
          "type" : "boolean"
        },
        "sender" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "receiver" : {
          "$ref" : "#/definitions/UserDTO"
        }
      }
    },
    "GameDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "year" : {
          "type" : "integer",
          "format" : "int32"
        },
        "image" : {
          "type" : "string"
        },
        "text" : {
          "type" : "string"
        }
      }
    },
    "GroupDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "owner" : {
          "$ref" : "#/definitions/UserDTO"
        }
      }
    },
    "UserDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "friends" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "$ref" : "#/definitions/UserDTO"
          }
        },
        "groups" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "$ref" : "#/definitions/GroupDTO"
          }
        }
      }
    }
  }
}